Fix CONST SINGLE generating wrong FFP literals (#76)
The CONST union stored singlenum as C float instead of LONG, corrupting
FFP bit patterns via implicit int-to-float conversion. Also fix several
places where C float/double literals (0.5, 10.0, -1.0) were passed to
SP* functions that expect LONG FFP values.

+++ src/ace/c/acedef.h
@@ -460,9 +460,9 @@ enum {	variable = 	3000,
 
 /* --symbol table-- */
 typedef union conststruct {
-			   SHORT shortnum;	
+			   SHORT shortnum;
 			   LONG  longnum;
-			   float singlenum;
+			   LONG  singlenum;
 			  } CONST;
 
 typedef struct structmem {
